fixed issues

+++ 0x0E-structures_typedef/2-print_dog.c
@@ -6,17 +6,17 @@
  *return: nothing if null
  */
 
- void print_dog(struct dog *d)
+void print_dog(struct dog *d)
 {
 
-	if(d == NULL)
+	if (d == NULL)
 	{
 
 		return;
 	}
 
 	printf("Name: %s\n", d->name ? d->name : ("nil"));
-	printf("Age: %f\n", d->age);
+	printf("Age: %f\n", (float) d->age);
 	printf("Owner: %s\n", d->owner ? d->owner : ("nil"));
 
 }
